deleted Course attributes made RaceCalculations class updated RaceController gave race scene an uncompleted leaderboard

+++ src/main/java/seng201/team0/Race.java
@@ -33,7 +33,7 @@ public class Race {
     }
 
 
-    public void updateRace(double timeElapsed) {
+    public void updateRace(int timeElapsed) {
         // Update the distance for each opponent
         for (OpponentCar opponent : opponents) {
             opponent.updateDistance(timeElapsed);
